# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamBox — Movie & Anime Streaming API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movies/trending:
    get:
      operationId: getMoviesTrending
      tags: [movies]
      summary: Get trending movies
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: Trending movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
  /movies/top:
    get:
      operationId: getMoviesTop
      tags: [movies]
      summary: Get top rated movies
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: genre
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Top movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
  /movies/search:
    get:
      operationId: searchMovies
      tags: [movies]
      summary: Search movies by title
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"
...
  /movies/{id}:
    get:
      operationId: getMovieById
      tags: [movies]
      summary: Get movie details by TMDB ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Movie detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /movies/{id}/embed:
    get:
      operationId: getMovieEmbed
      tags: [movies]
      summary: Get embed URL for a movie
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Embed info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedResponse"

  # ── WATCHLIST ───────────────────────────────────────────────────────────────
  /watchlist:
    get:
      operationId: getWatchlist
      tags: [watchlist]
      summary: Get user watchlist (session-based)
      responses:
        "200":
          description: Watchlist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WatchlistItem"
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add item to watchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchlistInput"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistItem"
  /watchlist/{itemId}:
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove item from watchlist
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ── Anime schemas ──
...
    EmbedResponse:
      type: object
      properties:
        embedUrl: { type: string }
        sources:
          type: array
          items:
            type: object
            properties:
              label: { type: string }
              url: { type: string }
            required: [label, url]
      required: [embedUrl, sources]

    # ── Movie schemas ──
    MovieItem:
      type: object
      properties:
        tmdb_id: { type: integer }
        title: { type: string }
        overview: { type: ["string", "null"] }
        poster: { type: ["string", "null"] }
        backdrop: { type: ["string", "null"] }
        rating: { type: ["number", "null"] }
        year: { type: ["integer", "null"] }
        genres: { type: array, items: { type: string } }
      required: [tmdb_id, title, genres]
    MovieDetail:
      type: object
      properties:
        tmdb_id: { type: integer }
        title: { type: string }
        overview: { type: ["string", "null"] }
        poster: { type: ["string", "null"] }
        backdrop: { type: ["string", "null"] }
        rating: { type: ["number", "null"] }
        year: { type: ["integer", "null"] }
        runtime: { type: ["integer", "null"] }
        genres: { type: array, items: { type: string } }
        cast: { type: array, items: { type: string } }
        trailer_url: { type: ["string", "null"] }
        imdb_id: { type: ["string", "null"] }
      required: [tmdb_id, title, genres, cast]
    MovieListResponse:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/MovieItem" } }
        total: { type: integer }
        page: { type: integer }
        hasNext: { type: boolean }
      required: [items, total, page, hasNext]

    # ── Genre ──
...
    WatchlistItem:
      type: object
      properties:
        id: { type: integer }
        contentId: { type: integer }
        contentType: { type: string, enum: [anime, movie] }
        title: { type: string }
        poster: { type: ["string", "null"] }
        score: { type: ["number", "null"] }
        addedAt: { type: string }
      required: [id, contentId, contentType, title, addedAt]
    WatchlistInput:
      type: object
      properties:
        contentId: { type: integer }
        contentType: { type: string, enum: [anime, movie] }
        title: { type: string }
        poster: { type: ["string", "null"] }
        score: { type: ["number", "null"] }
      required: [contentId, contentType, title]