# lib/api-zod/src/generated/types/movieItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface MovieItem {
  tmdb_id: number;
  title: string;
  /** @nullable */
  overview?: string | null;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  backdrop?: string | null;
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  year?: number | null;
  genres: string[];
}