# lib/api-zod/src/generated/types/watchlistInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistInputContentType } from './watchlistInputContentType';

export interface WatchlistInput {
  contentId: number;
  contentType: WatchlistInputContentType;
  title: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  score?: number | null;
}