# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamBox — Movie & Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
import type { WatchlistItemContentType } from './watchlistItemContentType';

export interface WatchlistItem {
  id: number;
  contentId: number;
  contentType: WatchlistItemContentType;
  title: string;
  /** @nullable */
  poster?: string | null;
  /** @nullable */
  score?: number | null;
  addedAt: string;
}